BF: limit CMD_MAX_ARG if obnoxious value is encountered.
Apparently the issue observed with elderly python 3.4 (no longer
supported) on Debian system, is likely to do nothing with the
Python but some peculiarity of the system since now was observed on
CentOS conda env with Python 3.9.

Since our hardcoded limits are quite generous, and 3.4 is no longer
support anyways, I decided just to add "flexible"  check that if
returned value is a million times larger (still have a buffer of
thousand times in this particular case) -- take hardcoded value.

Closes #5943