BF?(TODO Refine): encode command for execution
Apparently this solves nearly all failing tests on a system with non-utf8
default locale, but the question is it the right thing to do?

the only not "fixed" test is now failing differently test_run.py:test_py2_unicode_command

Decided to see what would be ramifications of such a change across
the board

Also harmonized and isort'ed imports.  Apparently tunes up spacing
for non-top-level imports as well