RF: Replace exc_str in result dicts
Makes use of an already existing approach to set an `exception` field
and have `get_status_dict` generate an `exception_traceback` field in
addition. Just like with logging this is intended to separate user
messaging from technical details.
Ultimately, it should probably be result renderers that pick up an
exception` field and display it (or not) in whatever way that renderer
sees fit. Hence passing the `CapturedException` instead of the original
exception, providing different formatting options to use for such a
mechanism seems like a better state for further redesign.
This patch adapts `get_status_dict` accordingly.

Also goes through all calls of exc_str, where it was used to build a
message in a dict (as opposed to composing a log message), in order to
account for that.