Change create_sibling_ghlike --auth setup
Now there is --credential to identify a credential to be retrieved
from any credential provider (config, env, store), instead of passing a
token on the command line.

This addresses the gist of
https://github.com/datalad/datalad/pull/5949#discussion_r701095470

This now also makes it possible to select a particular token out of a
set of multiple candidates for a specific site.