ENH: import datalad.api to bind extensions methods for discovery of dataset methods
This might be not the most elegant but the "easiest" in DX approach to resolve
this issue IMHO.  It might cause run time penalty though in small scripts which
import only a few dedicated modules, thus do not require import of the entire
API.

An alternative would be to e.g. do 2 rounds of that loop, importing datalad.api
only upon the end of the first loop if no meth is found.

Closes #5990