BF: Do not overwrite existing siblings with create-sibling-gitlab
In #6014 I reported that two create-sibling-gitlab calls with different repository names but
identical (default) sibling name overwrite each other - only the latest sibling is kept.
This behavior is confusing, and contrary to the behavior of all other create-sibling-* calls
that first check whether the remote name is already an existing sibling and error out when
existing isn't 'reconfigure' or 'replace'.
This change adds such a safe guard also for create-sibling-gitlab