BF: undeprecate readline_rstripped + adjust mock to mock actual warnings library
I was wrong in my statement that this function is no longer used, it is used in the Batched*
classes.  It is a trivial function and could be moved I guess to not appear in the interface
but since we do not really sanitize those, I decided that undeprecating would be the easiest
way forward.  Closes #6030.

Upon removal of "warnings" import test_done_deprecation started to fail since it patches
the local import of warnings. IMHO it should mock patch the actual warnings library, so I have
fixed that test accordingly