ENH/RIA: New parameter to only act on pre-existing stores
This implements a basic solution to #5982.
The issue is a feature request for the ability to only act on
create-sibling-ria calls if the specified store already
exists. This change adds a parameter to create-sibling-ria,
'--construct-store' (False by default), which, if set, creates a new
store if it doesn't exist under the specified url, and otherwise fails to
create a RIA sibling, warning the user that there is no store at the
specified path.

The parameter is intentially not named 'create-store' because this would
lead to a name conflict with the 'create_store' method of ria_utils.

TMP typos