Boost default performance of status/diff commands
This is done by removing the file-type evaluation as a default
operation, following the rational given in
https://github.com/datalad/datalad/issues/6084.

Moreover, the type reporting now follows a simpler rule.
When asked for a plain status, file-types are reported as known
to git. Only in `--annex` reporting mode the type info is
corrected for symlinks that are actually annexed files, and
the links only exist for technical reasons in a particular clone.

This generally leads to more robust and valid behavior across
supported environments, and to more performance, because this
kind of type inference can be done without expensive file system
interaction to resolve symlinks of read-out file content.

Using this simpler switching logic (--annex yes/no), also avoid
further expansion of the `diff` API to tune file-type evaluation
for diffs.

Fixes datalad/datalad#6084
Fixes datalad/datalad#6034
Fixes datalad/datalad#5643