WIP/RF: Transform add-archived-content into a dataset method
This is a very crude and work-in-progress refactoring effort for
add-archive-content. It doesn't yet deal with simplifying the code,
but only includes transitions to current dataset methods (e.g., proper
result rendering, a --dataset argument) and transitions away from things
that are so outdated or fragile that even I can spot them (e.g., use of
pathlib functionality instead of os.path functionality).