New method AnnexRepo.get_file_annexinfo()
This is the companion of get_content_annexinfo() for single file
queries. This is a common use case, not only in the tests.

This is (for now), not channeling all functionality that
get_content_annexinfo() provides (amending Git properties, etc),
but the API could be extended, if this turns out to be
useful in the future.

Importantly, with this method in use, in many places dedicated
additional calls to file_has_content() can now be reduced to a single
call to this new method.