BF: Prevent default recursion of create-sibling
By default, create-sibling was erroneously recursing into the first
level of subdatasets of a dataset, creating additional, unwanted
siblings.
(see #5614 for a corresponding issue)
This was due to a flawed logic with regard to --since:
--since by definition is only operating on subdatasets, but that
would require a recursive flag to be set. Yet even without a
recursive flag, create-sibling evaluated subdatasets states for
since, and with this, it added the first level of subdatasets
of any given dataset into the list of to-be-considered datasets
for sibling creation.

The change in this commit catches the use of since without a
correspoding recursive flag early on and raises a ValueError.
It also only evaluates state of subdatasets if there is any
need for recursion - if not, it simply just considers the
given dataset.