Clarify interaction between key and use-current-dir
When a --key is given, we do not extract the key in its current
location (the object tree), but prevent this from happening.
Without any other location to extract to, the command however fails,
unless --use-current-dir is given (reported in #3995).
The current docstring suggests, however, that use-current-dir has no
effect if used together with --key.

To clarify, this change explicitly enforces extraction in the current
directory (we don't know any other place), and states this in the
parameter docstring.