Replace AnnotatePaths in metadata commands
This is the second-to-last stop to finally fix #3368

There is only a minimal behavior change: A rather peculiar reporting
behavior, where metadata of the "nearest" available dataset is chosen
when no explicit dataset is given.

This has no practical value and would complicate this temporary fix
substantially. This whole stuff only needs to last until metalad is
fully functional.