BF: Force git-annex to commit the changes before we interrogate git-annex branch
The use case was detected while working on
https://github.com/datalad/datalad-crawler/pull/112
to accompany
https://github.com/datalad/datalad/pull/6105 .

Due to use of alwayscommit = False to minimize number of initial commits,
the repository ended up in a "peculiar" state with no useful commits
in git-annex branch -- all changes were staged within git-annex to be
committed "eventually".

Although we could have easily "fixed" the test in datalad-crawler,
I consider the approach of going around git-annex commands to interrogate
git-annex metadata to be fragile, in particular for such a case as here:
some other command could have been ran in alwayscommit=false mode, and
information in git-annex branch might not yet reflect actual state of
git-annex.  Ideally we should avoid doing such "going around git-annex".
As a quick resolution I have decided to just trigger a possible commit
by running  git-annex info --fast.  But may be there is a better/cheaper
way to trigger git-annex to commit, @joeyh?

Questionable change is the expansion of call_annex interface.  Could be
avoided through monkey patching of self.always_commit for that call,
but that would be more ugly IMHO