Change the behavior of run_procedure when called from create
This commit allows run_procedure to accept as input a dictionary
reported by run_procedure(discover=True), causing it to skip
calling _get_procedure_implementation().
This mechanism is used in create: first run_procedure(discover=True)
is executed to establish availability before creating the dataset,
and then the results are used to run the procedure on the created
dataset.