infra: Provide custom prefix to auto-related labels
This change is an attempt to act on the suggestion in #5833 to ease
understanding in how PRs need to be labelled for the 'auto' versioning
workflow. The test checking for appropriate labelling of PRs is called
'semver', hence my initial suggestion here in this change is a semver
prefix in hopes to make the relevant issues more discoverable to
new contributors.