Enable `@serve_path_via_http(tmp_dir, use_ssl=True)`
For this to work in a reliable way that needs no per-request fiddling,
we have to ship the entire CA/cert shebang.

A complete, pre-generated root CA, keys, and a 10-year certficiate+key
for 'localhost' URLs is shipped under `datalad/tests/ca`.

If it is detected, the internal HTTP server can be SSL-enable via the
new flag `use_ssl` (off by default).

On-startup the webserver is tested to have a functional SSL setup. If it
cannot be confirmed `SkipTest` is raised. Dedicated tests for
python-requests and stdlib-tooling are performed.

For a complete setup:

1) a root CA must be deployed

   # deploy certificate on Debian
   sudo mkdir -p /usr/local/share/ca-certificates/
   sudo cp datalad/tests/ca/ca-root.pem /usr/local/share/ca-certificates/datalad-root.crt
   sudo update-ca-certificates

2) python-requests must get the root CA in addition

   export REQUESTS_CA_BUNDLE="$(readlink -f datalad/tests/ca/ca_bundle.pem)"

   This is automatically done in `setup_package()`

Afterwards a

  `python -m nose -s -v datalad.tests.test_tests_utils:test_serve_path_via_http`

should confirm a working SSL setup.