Attempt to reimplement the datalad special remote with AnnexRemote
This is a likely incomplete draft. It primary aim is to determine what
is critical functionality and what is an artifact of the custom
annex protocol implementation.

Later changes will need to perform an similar process for the archive
special remote, and lastly both need to be consolidated to avoid
duplication.

I have removed the test that validated annex protocol compliance. We can
rely on AnnexRemote to do this properly.