Do not fail dropping an uninitialized AnnexRepo
Previously it would fail trying to determine whether the local
annex is known to remotes, but before `init` there is no local
annex.

With this change I am catching this condition and avoid the entire
investigation.

The potentially bigger issue (whether or not such a dataset should
carry an `AnnexRepo` instance, rather than `GitRepo`) is related
to #6103 and must be discussed separately.

Fixes datalad/datalad#6180