Distinguish application vs libary-type usage
This changeset introduces `datalad.enable_librarymode()` and the command
line switch `--library-mode` to put DataLad out of its normal
"application" runtime mode, and into a newly distinguished "library"
mode. In this new mode it shall, in the future, become a better citizen
in 3rd-party software ecosystems and applications.

A `datalad.in_librarymode()` function introduced to test for this
condition in the code base.

Fixes datalad/datalad#6081