TST: Mirror OSX approach to use GIT_CONFIG_GLOBAL on Windows
This change sits on top of gh-6251 and expands its change to
all platforms with Git version higher than 2.32. This should
ensure that we do not override the users USERPROFILE and HOME
variable by default and for all anymore (#6160). Instead, only
a small set of tests use this approach, and clean up after
themselves.