Enable ORA remote to recover from an interupted upload
Fixes datalad/datalad#6252 which comprises two problems.

ORA remote used the uuid of the *remote* to place a file in transfer
in a tmp dir. But this actually causes transfer conflicts with
simultanous uploads of the same key from multiple clones.

This is fixed by using the `here` uuid.

The second issue is a test for existence of the transfer file, and
failing if there is one already. This prevents the first issue from
having practical relevance re data integrity. With the usage of `here`
uuids, the conflict can no longer happen, and the test is obsolete.

The fix involves import and instantiation of `AnnexRepo()`. The
additional cost for that is minimal (~10-50ms on my machine) and is only
payed once per process at PREPARE. This newly available repo API, will
enable removal of custom code from the implementation in subsequent
changes.