ENH(TST)+RF: add test for consistency of having pos args and nonopts in CLI correspond
It includes adding explicit, and matching `args` for AddArchiveContent
and DownloadURL.  Without those this test would not be able to tell
them apart from other "peculiar" cases (such as added into signature
"filter_results" and alike; and rare but happening trailing options
given with nargs=REMAINDER).

Also, for that ad-hoc create-sibling-ria case, made `name` into a
non-kwarg since it is mandatory in CLI, so decided to unify.