OPT+RF: use @lru_cache for getargspec and get_sig_param_names
Those two functions are not really slow.  BUT for our means and purposes
we are not expecting signatures of the functions to mutate, so it makes sense
to avoid re-extraction of the signatures over and over again.  It brings runtime
for those functions from ~15mk, to 150ns - so about 100 times speed up on their
"execution"