Document that relative paths will not be made absolute
Resolving is now an explicit action. If it is explicitly off, and
relative paths are passed, the only way to turn them absolute is to
consider them relative to CWD -- this is by no means guaranteed to be
meaningful or intentional.

I now document the behavior. All internal usage is with absolute paths
anyways.