ConfigManager supports bare repos
When pointed to a bare repo, normal behavior for non-committed
config is maintained.

In addition, and replacing the readout of `.datalad/config` from
the worktree, `blob:HEAD:.datalad/config` is read (e.g., the
config committed in the default branch), which is the closest
analog of the worktree-config. Change detection for `reload()`
is also implemented, directly using the gitsha of this file.

Few minor issues are also contained in the changeset:

- missing parenthesis for function call, fixes datalad/datalad#6327

- wrong name for source config store, fixes datalad/datalad#6331

Alltogether, this fixes datalad/datalad#6264