BF: ORA over HTTP tried to check archive
Commit a4b87f0c58047c9337e0cd1604f2b3eaeaae99a7 broke this
by removing to much. CHECKPRESENT and TRANSFER-RETRIEVE would
try to address a potential archive if key couldn't be
found in annex object tree.

This is wrong. In opposition to remote execution variants
(SSH and local access), HTTP is supposed to only send one request
per key - archives have to be dealt with server-side.

Also adjust the ORA HTTP test to actually look into the results
of `fsck`, since a failure of hte special remote would not lead to
the call to `fsck` failing itself, but reporting in "note" or
"error-messages". The issue woud have been caught that way.

(Closes #6354)