BF: whereis key=True (non batched) needs to loop through entries "manually"
apparently no test was excercising it and no code was using where is
on multiple keys at once.  Detected this defect while working out the
"batch" mode for whereis

also adjusted the test for "conflicting" invocation of --all with
--key option.  Because we loop explicitly now, we would not even
invoke git-annex and thus not trigger the error if test is invoked
without any file.  So I just decided to adjust the test since IMHO it
would better reflect needed testing of us still bubbling up the
exception.

FWIW, filed an issue on --all not erroring out with --batch-keys:

https://git-annex.branchable.com/bugs/should_error_on_whereis_--batch-keys_--all/?updated