ENH: whereis key=True batch=True  mode support (requires recent git-annex)
Recently (in 8.20210903) git-annex introduced --batch-keys option to a
number of commands which otherwise had --batch mode, but not for
whenever specification was on keys instead of paths. I thought to make
use of whereis(..., key=True, batch=True) mode of operation to only
then discover that we do not have that supported.