ENH: git-credential-datalad breaks circle
It is possible to set up a cycle telling Git to query DataLad
for credentials and DataLad to query Git. While this seems a
highly questionable thing to do if both configs are made for a
specific URL, it is actually desirable to have DataLad's credential
helper be set up for use with Git unconditionally.
A DataLad provider config, that would then be set to query Git for a
particular (range of) URL(s), would end up calling git-credential,
calling git-credential-datalad, calling DataLad, calling git-credential,
calling git-credential-datalad, ...

However, from within the credential helper, we can actually detect that
and break the cycle. If we find `GitCredential` as the resut of querying
DataLad, we can simply stop.