RF: Deprecate annexstatus, move it from a class method into a test helper
This function was formerly used only in tests, with one exception in save
that relied on its precence as an attribute in ds.repo to determine
whether or not ds.repo was an AnnexRepo or GitRepo instance.

Tests that desperately rely on this, or test its function, have been
adjusted to use the test helper instead.

Work towards #6067