Make CLI autocompletion snappy
It basically inspects the incomplete argument list now, and
based on that can make more clever decisions what to load.

There are more possibilities for making things faster. For
example, we could load the extensions one by one, while searching
for unknown commands. But already now it feels great.

This change should have no impact on the full help generation. But it
breaks the interal operations needed to build manpages and docs.
This needs fixing, but was always ugly anyways.