RF+BF(TST): run "datalad " command externally to collect completions
oddly, even with patching both ._exit and .exit in prior version,
if multiple tests ran, test_completion would end the entire process
without given overall report etc.  So even though external invocation is
slower (takes nearly 10x longer for the test), it is more "robust"
for the test since we are not messing with the state of current
process.  Moreover it avoids getting that exception about inability
to close fd 9 as was with the previous "in process" implementation