OPT: create_sibling - diff to HEAD not None (tree)
Primary motivation is very slow --since= mode of operation of create-sibling
on a heavy superdataset (see #6399).  Although it could be argued that there might be
a benefit from create-sibling to be able to create siblings for not yet committed
submodules, I think such use cases are sparse if any and thus diff"ing to the committed
state (HEAD) is actually the desired effect.

Fixes #6399 (although I did not check explicitly)