Remove invalid AnnexCustomRemote.whereis() implementation
This was a leftover from before this class inherited from
`AnnexRemote.SpecialRemote`. In order to communicate that
a whereis isn't supported/meaningful, one would need to raise
`UnsupportedRequest()` instead -- which happens to be the default
implementation in the base class.

Therefore removal is the best and a sufficient fix.

This change does not include a proper test. Despite the protocol
violation, a `git-annex whereis` (the relevant call to test here) would
succeed. Consequently, the added test only verified that no
error messages pops up anymore.

Fixes datalad/datalad#6453