BF: No object tree in RIA w/o special remote
`create-sibling-ria` and the underlying helper created a dataset's
`annex/objects` directory unconditionally. However, only having the git
repositories in store with no special remote and no data whatsoever is
valid and supported usecase. The existence of that (empty) directory
increases the inode count per dataset, is superfluous and leads to
failures of annex commands being executed directly in store. While the
latter is something to generally be very careful about, there are valid
usecases and with the previous behavior we'd cause annex to pointlessly
fail on discovery of that dir, indicating to annex that the repo is
not annex-init'ed but was apparently used with annex before.