Set a minimum dependency on keyrings version 20.0 for token-auth in extensions
https://github.com/datalad/datalad-osf/issues/123 reported a failure for
token-based authentication when eldery versions of SecretStorage were used
that could not handle empty passwords (as is the case with token-based auth).
Updating the minimum dependency to keyring fixes this, because keyring versions
20.0 or higher have dependencies to minimal versions of SecretStorage that
contain a fix for this already. Now that this version of the keyring
package is widely available, we can incorporate this fix into datalad core.

fixes #6498 and https://github.com/datalad/datalad-osf/issues/123