OPT: create-sibling do not try to OPT by limiting paths to installed submodules
Prior approach did not  work well together well with --since specification, where
subdatasets recursively would traverse entire tree for no need. And it was added in
8c673a4d75f8d073a00dabfd00033f1744349ebe RFing as an OPT to avoid expensive traversal
of subdatasets which are not installed.  But for those there should be no extra cost
since we cannot descend into not installed subdatasets.  But may be there is cost since
we will be analyzing entire trees (I guess) and not just subdatasets, which is
what matters for create-sibling -- so there could be more OPT I think