Make SSH remote annex bundle sensing configurable and off by default
Previously, this feature was enabled unconditionally for any use of
`datalad sshrun`. We use this command as `GIT_SSH_COMMAND` and also
configure it to be used by git-annex. Consequently, it is used for
pretty much any SSH-related functionality.

This change introduces a new config switch to enable/disable this
feature, and the defaults of the underlying methods have been adjusted
to honor this switch.

As can be seen below, this has the potential to increase remote
execution performance substantially (here 500ms or 20+% for a call to
`uname`. The observed cost is per each call to `sshrun`.

```
===> multitime results
1: datalad -c datalad.ssh.try-use-annex-bundled-git=true sshrun judac.fz-juelich.de uname
            Mean        Std.Dev.    Min         Median      Max
real        2.530       0.126       2.367       2.526       2.703
user        0.266       0.015       0.253       0.263       0.294
sys         0.081       0.008       0.071       0.081       0.094

===> multitime results
1: datalad -c datalad.ssh.try-use-annex-bundled-git=false sshrun judac.fz-juelich.de uname
            Mean        Std.Dev.    Min         Median      Max
real        1.899       0.024       1.869       1.905       1.932
user        0.271       0.013       0.249       0.274       0.288
sys         0.064       0.018       0.041       0.064       0.093
```

The default of this config switch is OFF (in contrast to the previous
default behavior). Rational: In general it is nohow guaranteed that

- there will be a git-annex installation on the remote end there
- a potentially existing bundled Git is "better" (could be missing
  system-provided security fixes)

Fixes datalad/datalad#6530