UX: Better message when ORA does not find store via HTTP
This let's ORA account for HTTP failures during initremote/enableremote
and report an inaccessible store (and its URL).

Otherwise we'd spit out a possibly gigantic exception traceback that is
not helpful for users. With support for annexremote's logging framework
we'll have other means for detailed messaging soon.

The approach implemented here tries to have a somewhat uniform handling
across different IO classes by raising FileNotFoundError from within
`HTTPRemoteIO.read_file` if we encountered a 404, while all other
exceptions coming from the downloaders are bubbling up as they indicate
a more general issue.
AccessDeniedError and AccessFailedError are handled by the special
remote class to provide a leaner, user-facing message when this happens
during the verification of the store itself or the dataset within.
Everything else will lead to failure as before.