Stop forcefully disabling keyboard-interactive SSH auth
We did this initially for non-multiplex connections, assuming that it
would not work "by-design" when the SSH process is controlled by a
process that is not directly connected to a terminal on windows.

However, explorations documented in https://github.com/datalad/datalad/issues/5829
and https://github.com/datalad/datalad/issues/6524 revealed that this is
wrong, and the reason for it appearing to not work is in the protocol
handling within `Runner`.

This patch merely removed the forceful disabling, and thereby
immediately makes `sshrun` functional on windows.