TST: Enable codecov coverage for subprocesses
The changes follow the documentation in
coverage.readthedocs.io/subprocess.html and
pymotw.com/2/site/#sitecustomize. They are based on a config variable
switch to enable subprocess coverage reporting, and the introduction of
a sitecustomize.py script, that performs a coverage process startup for
non-datalad processes.

I encountered unexpected behavior in that a system-wide installation of
DataLad interfered with coverage reports - I suspect this is because I
altered the PYTHONPATH variable to make the sitecustomize.py script
discoverable. This unexpected behavior manifested in coverage debug
messages stating that DataLad and other packages installed in
site-packages would be part of stdlib, which coverage by default would
not trace. I worked around this by changing datalad's installation to a
user-based installation. Maybe this could also be resolved by placing
sitecustomize.py into a path that already is in the Pythonpath (Lib/),
but I did not go down this route.

Fixes #6367