New credentials() command and CredentialManager
This is a new subsystem aiming to factor out the management of
credentials from the downloaders/providers concept into a
standalone system.

Importantly, the management of downloaders/providers credentials is
supported by the new system, albeit requiring "type hints" for
some operations, because downloader/provider credentials are only
fully discoverable in the context of a provider (identified by the
context of a request).

In contrast, the new system considers credentials as a standalone,
request independent entity.

A credential is discoverable for any number of abitrarily name
properties, but is typically (and right now exclusively) identified
by a name.

Two new config settings are introduced

- datalad.credentials.repeat-secret-entry
- datalad.credentials.hidden-secret-entry

that control who secrets are entered in interactive sessions.

The system of provided operations/actions enables the implementation
of a confirm-function-function-before-storage pattern (see next commit,
ping datalad/datalad#3126).

Fixes datalad/datalad#396
Fixes datalad/datalad#2721
Fixes datalad/datalad#6519