Make datalad.cli.tests.test_main.run_main() also capture UI output
This is necessary, because the UI holds its own handle of `sys.stdout`
that was not affected by the previous patch approach.

With this change, the same StringIO instance is used for both
`sys.stdout` and the `UI.message()` outlet.

With this change, `run_main()` is no suitable for testing more CLI
aspects, such as result renderer output in response to CLI calls.

The resulting three context manager calls have been collapsed into
a single statement, rather than continuing the pattern from before.

Closes datalad/datalad#6561