Deduplicate top-level CLI exception handler output
Whenever an exception reached the `main()` exception handler, it is
converted to a `CapturedException()` and `str(CapturedException)`,
plus the exception type is logged. However, the exception type
is already contained in the `__str__` output:

```
[ERROR  ] ValueError(Value or unset flag ":" missing for property [('ddd',)]) (ValueError)
```

With this change, this duplication is removed and changes the output to

```
[ERROR  ] Value or unset flag ":" missing for property [('ddd',)] (ValueError)
```

This is specifically not identical to the output of
`str(CapturedException())`, in order to maintain the previous spirit of
putting more emphasis on the user-facing message, rather than the
technicality of the specific exception name.

Fixes datalad/datalad#6549