Prevent internal download functionality from logging at INFO level
See #6545 for the rational. #6545 was filed as compound issue, hence
cannot be solved here.

Logging is maintained, but at DEBUG level. This enables using the
downloader functionality when the download is a technical detail (fetch
remote state info) rather then the primary goal (as with data downloads)
-- without unconditionally polluting user-facing output with traced of
internal processing.

It is unclear for which functionality the INFO level message would be or
was possibly essential, but any such functionality should either use
progress logging, or result reporting nowadays.