RF(possible OPT): store common all_deleted before analyzing them
Could be of great benefit whenever deletions are only few
among many other changes.  This way we would not need to go
through all N twice, and only once.  If everything is removals,
then we would actually pay the penalty, but assuming that we saving
more often additions and changes and not removals -- we would
gain improvement