BF+RF: do not use "status" where we do not remove "clean" entries
Then we were providing such status  even with "clean" entires to _save_post
leading to e.g. asking to commit empty directories etc which caused tests
failures.

Instead of "filtering out" status dict, I decided to just stop using it
completely and RF _save_post to not be provided with status dict, but with
what it actually needs -- list of files.  And for that list, we just now
create it from groupped dicts, instead of original singular status