Make sure a download failure reason is always communicated
Previously the datalad special remote would report the `__cause__`
of an exception, if there was one set -- leading to uninformative
error messages of the type:

```
  Failed to download from any of 1 locations [None]
```

for any exception without a declared `__cause__`.

This change falls back on the corresponding `CapturedException` for
reporting, whenever no `__cause__` is available. This avoid non-specific
error messages.