Channel progress logging by AnnexSpecialRemoteProgressBar to git-annex
(again)

This was removed (unintentionally) when the datalad and -archives
special remotes were reimplemented in
https://github.com/datalad/datalad/pull/6165

I only now understood (for a brief moment in time) how this actually
works.

This change places a dedicated method in the AnnexCustomRemote
base class (again) that serves as a receiver and place of
documentation.

Fixes datalad/datalad#6572