Consolidate entrypoint processing on importlib.metadata
This approach is now an offiical part of the standard lib (since 3.10),
and is available for earlier version in a provisional module and
also a separate package.

This change introduces `datalad.support.entrypoints` with
`iter_entrypoints()` that is now used as a uniform implementation.

As previously shown in https://github.com/datalad/datalad/issues/6584
this brings significant performance improvements for internal
processing. Moreover, they translate to outward-facing speed gains.
On a system with 10 extensions installed a `datalad -h` now completes
in ~300ms, whereas current mainline after the CLI update needs ~650ms,
and the latest release 0.15.6 needs ~1200ms -- so overall a 4x speed-up.

Importantly, the new `iter_entrypoints()` helper does not return complex
entrypoint instances anymore, which previously contaminated DataLad code
with the internals of `pkg_resources`. This is also the reason for one
occurrence of a `pkg_resources`-based entrypoint iteration being left in
the code (in `search` and its tests). It specifically relies on these
internals and even mocks them in the tests.

This code is likely to change entirely soon, and is also slow enough to
not make the require effort to switch it over felt by users.

This change implies no necessary modifications of any extensions.

Fixes datalad/datalad#6584
Fixes datalad/datalad#6585