RF: set allowed-url/http in test HOME .git/config instead of config calls per instance
This is inspired as a possible workaround for the
https://github.com/datalad/datalad/issues/6482
which is probably sufferring from another issue, but looking
at this code, I do not see why just not to set it in the fixture.

The only cons I can foresee is the test scenario where in fixture
we do not prepare our home whenever GIT_HOME is set.  IIRC it
was the case in debian package building case, but if needed,
we could set those up in debian/rules