Fix for annex' reporting on unknown paths
As of annex' 10.20220127 an issue became apparent (actually older),
where the reporting of annex commands with respect to paths that are
unknown to git or not existent at all happens in different ways. This
depends on annex' internal git calls (git itself appears to be
inconsistent WRT this message) as well as on the value of
'annex.skipunknown' the default of which changed in said version. This
results in both styles of reporting being used and as long as we rely on
parsing that output, we need to account for both.
Moreover, annex commands don't always exit non-zero when that happens,
which is why we need to check stderr output when the command doesn't
fail, too.
Fixed annex (ce91f10132805d11448896304821b0aa9c6d9845,
Feb 28, 2022 - "fix annex.skipunknown false error propagation"), which
is part of version 10.20220322.

None of that actually depends on annex' version exactly. It's mostly
just that the zero-exit bug in combination with the switch of default
made those issues surface.

(Closes #6492)
(Closes #6571)