BF/RF: create-sibling follow diff with constant_refs=False + operate on committed
Initial motivation for this change was fixing

   https://github.com/datalad/datalad/issues/6596

where while specifying --since, create-sibling  did not detect needing to
create a sibling for a subdataset.  A stale (missed during #6528 OPT/RF)
comment said that value of constant_refs=True should not matter since it was
not recursive originally. But now here we do recurse and it does matter since
we migth completely skip subtree where such refs are not yet known.  Changing
to False resulted in the test failing which made me realize that sibling was
created even for not yet committed changes to submodules.  Since for push it
would really not work to push something which is not yet saved, I think the
change in behavior for create-sibling is also desired.  And it fixes #6596.
So I have "fixed" and expanded test with the problematic case I had
discovered.